+++ src/Exceptions/templates/error-file.php
@@ -1,7 +1,8 @@
+<title>Solital error</title>
 <body style='background: #F8F8FF;'>
     <div style='background: #FFF; padding: 15px; font-family: sans-serif;'>
         <p>
-            <h1 style='color: #EE0000;'>FileSystem alert: error in code execution</h1>
+            <h1 style='color: #EE0000;'>Solital alert: error in code execution</h1>
         </p>
         <hr>
         <p><strong>Type error:</strong> Not Found<br>