+++ src/Http/Exceptions/InvalidArgumentException.php
@@ -8,7 +8,7 @@ class InvalidArgumentException extends HttpException
 {
     public static function alertMessage(int $code, string $msg) 
     {
-        include_once ROOT.'/app/Core/Exceptions/templates/error-http.php';
+        include_once dirname(__DIR__, 2).'/Exceptions/templates/error-router.php';
         die;
     }
 }