+++ src/Http/Security/TokenProviderInterface.php
@@ -4,19 +4,13 @@ namespace Solital\Core\Http\Security;
 
 interface TokenProviderInterface
 {
-
-    /**
-     * Refresh existing token
-     */
-    public function refresh(): void;
-
     /**
      * Validate valid CSRF token
      *
      * @param string $token
      * @return bool
      */
-    public function validate(string $token): bool;
+    public function validate(): bool;
 
     /**
      * Get token token
@@ -25,5 +19,12 @@ interface TokenProviderInterface
      * @return string|null
      */
     public function getToken(?string $defaultValue = null): ?string;
+
+    /**
+     * Set csrf token
+     * @param int $seconds
+     * @return string
+     */
+    public function setToken(int $seconds = 20): string;
 
 }