+++ src/Resource/Session.php
@@ -8,8 +8,13 @@ class Session
      * @param string $index
      * @param mixed $value
      */
-    public static function new(string $index, $value): object
+    public static function new(string $index, $value, string $key = null): object
     {
+        if ($key != null) {
+            $_SESSION[$index][$key] = (is_array($value) ? (object)$value : $value);
+            return new static;
+        }
+
         $_SESSION[$index] = (is_array($value) ? (object)$value : $value);
         return new static;
     }
@@ -18,19 +23,34 @@ class Session
      * @param string $index
      * @return bool
      */
-    public static function delete(string $index): bool
+    public static function delete(string $index, string $key = null): bool
     {
+        if ($key != null) {
+            if (isset($_SESSION[$index][$key])) {
+                unset($_SESSION[$index][$key]);
+                return true;
+            }
+        }
+
         if (isset($_SESSION[$index])) {
             unset($_SESSION[$index]);
             return true;
         }
+
+        return false;
     }
     
     /**
      * @param string $index
      */
-    public static function show(string $index)
+    public static function show(string $index, string $key = null)
     {
+        if ($key != null) {
+            if (isset($_SESSION[$index][$key])) {
+                return $_SESSION[$index][$key];
+            }
+        }
+
         if (isset($_SESSION[$index])) {
             return $_SESSION[$index];
         }
@@ -38,9 +58,18 @@ class Session
 
     /**
      * @param string $index
+     * @return bool
      */
-    public static function has(string $index): bool
+    public static function has(string $index, string $key = null): bool
     {
+        if ($key != null) {
+            if (isset($_SESSION[$index][$key])) {
+                return true;
+            } else {
+                return false;
+            }
+        }
+
         if (isset($_SESSION[$index])) {
             return true;
         } else {